will run either the class files or the jar

+++ volacal.sh
@@ -1,1 +1,11 @@
-java -cp classes/ Main
+#!/bin/bash
+
+if [ -e volacal.jar ]; then
+	echo "Using jar file"
+	java -jar volacal.jar
+elif [ -e classes/Main.class ]; then
+	echo "Using class files directly"
+	java -cp classes/ Main
+else 
+	echo "Nothing to run"
+fi
